feat(add-bottles): refactoring add bottles, add friend picker component, allow gifting to multiple friends
* wip(add-bottle): try

* feat: convert pick friend to list adapter

* feat: convert pick friend to list adapter

* chore: split wine repository

* chore: move file to domain package

* chore: wip refactoring repositories

* feat: refact end

* fix(add-bottle): bind add friends methods

* chore: repositories refactoring

* fix: prevent date picker memory leak

* feat: wip data fetching with FriendPickerView

* feat: make add bottle friend appear on history

* feat: add signle friend closeable behaviorry

* fix: wip bottom sheet

* fix: fix search backdrop not toggling on animated checkbox click

* feat: switch back to dialog option

* feat: improve other info fragment ui

* feat(friend-picker): add filter & sort wip

* fix(friend-picker): fix adapter not picking up close icon changes

* fix(friend-picker): fix checkbox click not selecting friend

* feat(friend-picker): revert back to bottom sheet implementation

* feat(friend-picker): add scroll to friend

* fix(friend-picker): resolve monkey crash

* fix(friend-picker): fix bottom sheet multiplying on configuration change

* fix(friend-picker): fix ViewGroup LayoutTransition conflict with ViewPager

* feat(friend-picker): rework friend picker

* feat(friend-picker): add text switcher

* feat(friend-picker): add friend picket to add tasting screen

* feat(bottle-details): reword edit bottle

* chore: update privacy policy with app permission details

* fix(friend-picker): fix a crash when openin a bottle that was not given by someone

* fix(home): restore drawer cut corners

* feat(friend-picker): improve image finding

* chore: format

* chore: code cleanup

* chore: code cleanup

* chore: code cleanup

+++ app/src/main/java/com/louis/app/cavity/ui/search/filters/WineFilters.kt
@@ -1,6 +1,7 @@
 package com.louis.app.cavity.ui.search.filters
 
 import com.louis.app.cavity.db.dao.BoundedBottle
+import com.louis.app.cavity.domain.history.toInt
 import com.louis.app.cavity.model.*
 import com.louis.app.cavity.util.toBoolean
 import com.louis.app.cavity.util.toInt
@@ -112,7 +113,7 @@ class FilterFriend(private val friendId: Long, private val historyEntryType: Int
     override fun meetFilters(boundedBottle: List<BoundedBottle>): List<BoundedBottle> {
         return boundedBottle.filter {
             return@filter it.historyEntriesWithFriends.find { entryWithFriends ->
-                entryWithFriends.historyEntry.type == historyEntryType &&
+                entryWithFriends.historyEntry.type.toInt() == historyEntryType &&
                         (friendId in entryWithFriends.friends.map { f -> f.id })
             } != null
         }
